Refactor integrator modules; rename _protocol.py to _integrator_protocol.py and update imports accordingly
