+++ braincell/_integrator_implicit.py
@@ -25,7 +25,7 @@ from ._integrator_exp_euler import _exponential_euler
 from ._integrator_runge_kutta import rk4_step
 from ._integrator_util import apply_standard_solver_step, jacrev_last_dim
 from ._misc import set_module_as
-from ._protocol import DiffEqModule
+from ._integrator_protocol import DiffEqModule
 from ._typing import T, DT
 
 __all__ = [
