+++ braincell/_integrator_runge_kutta.py
@@ -21,7 +21,7 @@ import brainunit as u
 import jax
 
 from ._misc import set_module_as
-from ._protocol import DiffEqState, DiffEqModule
+from ._integrator_protocol import DiffEqState, DiffEqModule
 from ._typing import T, DT
 
 __all__ = [
