+++ braincell/_integrator_staggered.py
@@ -19,7 +19,7 @@ import brainunit as u
 from ._integrator_exp_euler import ind_exp_euler_step
 from ._integrator_voltage_solver import dhs_voltage_step
 from ._misc import set_module_as
-from ._protocol import DiffEqState, DiffEqModule
+from ._integrator_protocol import DiffEqState, DiffEqModule
 
 __all__ = [
     'staggered_step',
