+++ braincell/_multi_compartment.py
@@ -21,7 +21,7 @@ import brainunit as u
 from ._base import HHTypedNeuron, IonChannel
 from ._integrator import get_integrator
 from ._morphology import Morphology
-from ._protocol import DiffEqState
+from ._integrator_protocol import DiffEqState
 from ._typing import Initializer
 
 __all__ = [
