+++ braincell/_single_compartment.py
@@ -20,7 +20,7 @@ import brainunit as u
 
 from ._base import HHTypedNeuron, IonChannel
 from ._integrator import get_integrator
-from ._protocol import DiffEqState
+from ._integrator_protocol import DiffEqState
 from ._typing import Initializer
 
 __all__ = [
