+++ braincell/channel/hyperpolarization_activated.py
@@ -10,7 +10,7 @@ import brainstate
 import brainunit as u
 
 from braincell._base import Channel, HHTypedNeuron
-from braincell._protocol import DiffEqState
+from braincell._integrator_protocol import DiffEqState
 
 __all__ = [
     'Ih_HM1992',
